Several changes to include internationalization
i18n configuration is added in `astro.config.mjs` to
detect pages in english like `/en/now`.

Index page and takeoff mdx were added to test locale
 in english.

Components were affected to receive language and
display their content in English or Spanish.

Template page `[...slug].astro` was modified to
handle default slug and avoid issues when different
mdx files has same name (but location in folder
structure is different).

A fix was implemented in dates inside of post,
because date was being displayed as one day before,
in other words if a mdx has as date Feb 14th, then
site was displaying that date post as Feb 13th. With
this fix posts are displaying dates normally.

PostLayout was deleted, I think that it is not
necessary.