Fixed date on MDX files and post collection issue
Publication date on posts now contains time and is
displayed on pages and components.

Also, the warning: `The posts collection is defined
but no content/posts folder exists in the content
directory. Create a new folder for the collection,
or check your content configuration file for typos.`
is fixed. I had to rename `post` folder to `posts`
and I did some modifications where content
colllection is used.