RSS improvements
A new function was added to clean and format post's bodies, also this function is placed in a separate file for better readibility.
Markdown plugin was used and configured again to convert markdown syntax into html code.

Some posts were modified to use markdown syntax.

Assets folder and its content was deleted because they are not necessary now.